revision1
revisions to the design of site. updated the home page, and updated the "myExpertise" page.  added some additional css that I wasn't able to see. changed "index.php" file to "home.html". Couldn't open page in chrome for some reason. Helped me to decipher what was what without the code breaking on the site. Transfer this theme to the other pages in terms of the nav bar, and maybe having a footer on each of the pages for consistency. 

Also worked on the bio. 

MyExpertise page has icons that look "smooshed". May need to work on opening those up to look more professional. 

Icons on Contact page need labels. One for "email" one for "linkedIn" and one for "Resume" so folks know what icon is what.